backend/requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LLM Integration
langchain>=0.0.335
openai>=1.3.0
llama-index>=0.9.0
tiktoken>=0.5.1

# Vector Database
chromadb>=0.4.18

# Code Processing
tree-sitter>=0.20.1
pygments>=2.16.1

# Testing
pytest>=7.4.3
httpx>=0.25.0

# Utilities
numpy>=1.26.1
pandas>=2.1.1